% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locs.R
\name{check_deployment}
\alias{check_deployment}
\title{Check deployment}
\usage{
check_deployment(DT, meta)
}
\arguments{
\item{meta}{metadata with variable 'deployment' indicating path to csv with three columns indicating 'id' animal id, 'start_date' start of deployment and 'end_date' end of deployment both structured as 'YYYY-MM-DD' formatted character}
}
\value{

}
\description{
Check deployment
}
