% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locs.R
\name{check_locs}
\alias{check_locs}
\title{Check locs}
\usage{
check_locs(DT, meta)
}
\arguments{
\item{meta}{NA}
}
\value{

}
\description{
using functions check_longlat and check_loc_meta
}
\author{
Alec L. Robitaille
}
