% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locs.R
\name{check_locs_meta}
\alias{check_locs_meta}
\title{Check loc meta}
\usage{
check_locs_meta(DT)
}
\arguments{
\item{DT}{data.table}
}
\value{
data.table
}
\description{
Check location data's meta information eg. map quality, DOP, manual exclusion
}
\author{
Alec L. Robitaille
}
