% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locs.R
\name{check_longlat}
\alias{check_longlat}
\title{Check long and lat}
\usage{
check_longlat(DT)
}
\arguments{
\item{DT}{data.table}
}
\value{
data.table
}
\description{
Check long and lat
}
\author{
Alec L. Robitaille
}
