% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_csv.R
\name{export_csv}
\alias{export_csv}
\title{Export CSVs}
\usage{
export_csv(DT, outpath, splitBy, extracols)
}
\arguments{
\item{splitBy}{}
}
\value{

}
\description{
Minimal wrapper around fwrite, to write separate CSVs if splitBy is provided
}
\author{
Alec L. Robitaille
}
