% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Generate data details}
\usage{
metadata()
}
\arguments{
\item{path}{filepath}

\item{x_long}{unprojected longitude coordinate column name}

\item{y_lat}{unprojected latitude coordinate column name}

\item{id}{individual identifier column name}

\item{datetime}{datetime column name. either provide datetime or both date and time}

\item{date}{date column name. either provide both date and time or datetime}

\item{time}{time column name. either provide both date and time or datetime}

\item{extracols}{list of extra column names to preserve}

\item{tz}{time zone of datetime columns as stored in the input file}

\item{epsgin}{EPSG code of input. if numeric, EPSG code to use for all rows. if character, column name of EPSG code in DT}

\item{epsgout}{local projection EPSG code to reproject x_long+y_lat into}

\item{deployment}{path to deployment CSV. see \code{check_deployment()}.}
}
\value{

}
\description{
Metadata about input data including file paths, column names, time zones and EPSG codes.
}
\author{
Alec L. Robitaille
}
