% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_locs.R
\name{project_locs}
\alias{project_locs}
\title{Project locs}
\usage{
project_locs(DT, epsgin, epsgout)
}
\arguments{
\item{EPSG}{numeric; EPSG code}
}
\value{

}
\description{
Project locs
}
\author{
Alec L. Robitaille
}
