% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(path, meta)
}
\arguments{
\item{path}{character; path to input file passed to fread}
}
\value{
data.table
}
\description{
Read data
}
\examples{
path <- system.file('extdata', 'DT.csv', package = 'spatsoc')
DT <- read_data(path = path)
}
\author{
Alec L. Robitaille
}
