% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_colnames.R
\name{set_colnames}
\alias{set_colnames}
\title{Set consistent column names}
\usage{
set_colnames(
  DT,
  x_long,
  y_lat,
  id,
  date = NULL,
  time = NULL,
  datetime = NULL,
  extracols = NULL
)
}
\arguments{
\item{extracols}{character vector; list of column names to include (eg. for group_pts splitBy)}
}
\value{
Input DT with only selected columns id, datetime, coords and any column provided in extracols
}
\description{
Set consistent column names
}
\author{
Alec L. Robitaille
}
